pyproject.toml
[tool.poetry]
name = "bank_api"
version = "0.1.0"
description = "A bank service project using FastAPI, PostgreSQL, and Redis"
authors = ["cyan92128505 <cyan92128505@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12.4"
annotated-types = "0.7.0"
anyio = "4.6.2.post1"
asyncpg = "0.30.0"
fastapi = "0.115.2"
greenlet = "3.1.1"
passlib = "1.7.4"
psycopg2-binary = "2.9.10"
pydantic = "2.9.2"
pydantic-settings = "2.6.0"
python-dotenv = "1.0.1"
python-jose = "3.3.0"
redis = "5.1.1"
SQLAlchemy = "2.0.36"
uvicorn = "0.32.0"

[tool.poetry.dev-dependencies]
pytest = "8.3.3"
pytest-asyncio = "0.24.0"
httpx = "0.27.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning:passlib.*:",
    "ignore::DeprecationWarning:jose.*:",
]